+++ utils/text/transformers.py
@@ -166,3 +166,18 @@ def seo_tokenize(title, lower=True, preserve_ascii_extended=False, preserve_unic
     tokenized_title = '-'.join(cleaned_title.split())
 
     return tokenized_title
+
+
+def snake_case_to_camel_case(string):
+    """Convert `snake_case` string to `CamelCase`"""
+    camel_string = ''.join(
+        part.capitalize() for part in string.lower().split('_')
+    )
+    return camel_string
+
+
+def snake_case_to_lower_camel_case(string):
+    """Convert `snake_case` string to `camelCase`"""
+    camel_string = snake_case_to_camel_case(string)
+    camel_string = camel_string[0].lower() + camel_string[1:]
+    return camel_string
